// frontend/src/shared/api/generated/candidates/models/fspInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Candidates
 * OpenAPI spec version: 0.1.0
 */
import type { FspAchievementResponse } from './fspAchievementResponse';

export interface FspInfo {
  linked: boolean;
  participant_id: string | null;
  synced_at: string | null;
  achievements: FspAchievementResponse[];
}